Moved widget files to src/widgets folder
