+++ main.cpp
@@ -1,4 +1,4 @@
-#include "src/gamewidget.hpp"
+#include "src/widgets/gamewidget.hpp"
 #include <QApplication>
 
 int main(int argc, char *argv[])
