+++ src/widgets/gamewidget.cpp
