+++ src/widgets/mazemapwidget.cpp
