+++ src/widgets/mazemapwidget.hpp
